requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
python-dotenv==0.17.1
itsdangerous==2.0.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
Werkzeug==0.16.1
SQLAlchemy<1.4.0
PyJWT==1.7.1